flare_backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
supabase>=2.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
dj-database-url>=2.1.0
channels>=4.0.0
channels-redis>=4.1.0
